feat: refactor conflict feature reading into a separate utility function
